requirements.txt
boto3==1.24.89
faker
Flask
pynamodb
python-dotenv